+++ index.ts
@@ -166,6 +166,8 @@ export const registerMenu = (menuId: string): boolean => {
 // @ts-ignore
 export { default as IonTabs } from "./components/IonTabs.svelte";
 // @ts-ignore
+export { default as IonTabsLegacy } from "./components/IonTabsLegacy.svelte";
+// @ts-ignore
 export { default as IonPage } from "./components/IonPage.svelte";
 // @ts-ignore
 export { default as IonBackButton } from "./components/IonBackButton.svelte";